requirements.txt
#####python版本 3.13.5
# IT资产生命周期管理系统 (ALMS) - 依赖包
# FastAPI 核心框架
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.1
orjson>=3.10.0  # 高性能JSON库，正确处理UTF-8编码

# 数据库相关
SQLAlchemy==2.0.41
PyMySQL==1.1.1
greenlet==3.2.3

# 数据验证和序列化
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
annotated-types==0.7.0

# 认证和安全
passlib==1.7.4
bcrypt==4.0.1

# 工具库
python-dotenv==1.1.1
click==8.2.1
colorama==0.4.6
typing_extensions==4.14.1
typing-inspection==0.4.1

# 文件处理
pandas==2.1.4
openpyxl==3.1.2
python-multipart==0.0.6

# 网络和异步
anyio==4.9.0
h11==0.14.0
sniffio==1.3.1
idna==3.10
dnspython==2.7.0
email_validator==2.2.0

# HTTP客户端（用于调用外部API）
httpx==0.24.1

# 日志系统
python-json-logger==2.0.7
python-logstash-async==2.5.0
elasticsearch==8.11.0

# 服务注册与配置中心
nacos-sdk-python==0.1.17

# 可选：开发和测试工具
# pytest==7.4.0
# pytest-asyncio==0.21.1

